+++ test/cli.js
@@ -20,21 +20,13 @@ describe('cli', function () {
             parse: function (commands, dir) {
               checks.bag_parse_commands = commands;
               checks.bag_parse_dir = dir;
-            },
-            readCustomConfigFileSync: function (file) {
-              checks.bag_configFile = file;
-              if (mocks.bag_configFile) {
-                return mocks.bag_configFile;
-              } else {
-                throw new Error('Config file not found.');
-              }
             }
           }
         },
-        './couchpenter': function (url, config, dir) {
+        './couchpenter': function (url, configFile, opts) {
           checks.couchpenter_url = url;
-          checks.couchpenter_config = config;
-          checks.couchpenter_dir = dir;
+          checks.couchpenter_configFile = configFile;
+          checks.couchpenter_opts = opts;
           return {
             config: function (exit) {
               checks.couchpenter_config_exit = exit;
@@ -74,7 +66,6 @@ describe('cli', function () {
     });
 
     it('should contain setup command and delegate to couchpenter do when exec is called', function () {
-      mocks.bag_configFile = '{ "foo": "bar" }';
       checks.bag_parse_commands.setup.desc.should.equal('Create databases, then create/update documents');
       checks.bag_parse_commands.setup.options.length.should.equal(3);
       checks.bag_parse_commands.setup.action({
@@ -83,8 +74,8 @@ describe('cli', function () {
         dir: 'curr/dir/'
       });
       checks.couchpenter_url.should.equal('http://localhost:5984/somedb');
-      checks.couchpenter_dir.should.equal('curr/dir/');
-      checks.couchpenter_config.foo.should.equal('bar');
+      checks.couchpenter_opts.dir.should.equal('curr/dir/');
+      checks.couchpenter_configFile.should.equal('someconfigfile.js');
       checks.couchpenter_do_tasks.length.should.equal(2);
       checks.couchpenter_do_tasks[0].should.equal('setUpDatabases');
       checks.couchpenter_do_tasks[1].should.equal('setUpDocuments');
@@ -92,7 +83,6 @@ describe('cli', function () {
     });
 
     it('should contain setup-db command and delegate to couchpenter do when exec is called', function () {
-      mocks.bag_configFile = '{ "foo": "bar" }';
       checks.bag_parse_commands['setup-db'].desc.should.equal('Create databases only');
       checks.bag_parse_commands['setup-db'].options.length.should.equal(3);
       checks.bag_parse_commands['setup-db'].action({
@@ -101,15 +91,14 @@ describe('cli', function () {
         dir: 'curr/dir/'
       });
       checks.couchpenter_url.should.equal('http://localhost:5984/somedb');
-      checks.couchpenter_dir.should.equal('curr/dir/');
-      checks.couchpenter_config.foo.should.equal('bar');
+      checks.couchpenter_opts.dir.should.equal('curr/dir/');
+      checks.couchpenter_configFile.should.equal('someconfigfile.js');
       checks.couchpenter_do_tasks.length.should.equal(1);
       checks.couchpenter_do_tasks[0].should.equal('setUpDatabases');
       checks.couchpenter_do_exit.should.be.a('function');
     });
 
     it('should contain setup-doc command and delegate to couchpenter do when exec is called', function () {
-      mocks.bag_configFile = '{ "foo": "bar" }';
       checks.bag_parse_commands['setup-doc'].desc.should.equal('Create documents only');
       checks.bag_parse_commands['setup-doc'].options.length.should.equal(3);
       checks.bag_parse_commands['setup-doc'].action({
@@ -118,15 +107,14 @@ describe('cli', function () {
         dir: 'curr/dir/'
       });
       checks.couchpenter_url.should.equal('http://localhost:5984/somedb');
-      checks.couchpenter_dir.should.equal('curr/dir/');
-      checks.couchpenter_config.foo.should.equal('bar');
+      checks.couchpenter_opts.dir.should.equal('curr/dir/');
+      checks.couchpenter_configFile.should.equal('someconfigfile.js');
       checks.couchpenter_do_tasks.length.should.equal(1);
       checks.couchpenter_do_tasks[0].should.equal('setUpDocuments');
       checks.couchpenter_do_exit.should.be.a('function');
     });
 
     it('should contain teardown command and delegate to couchpenter do when exec is called', function () {
-      mocks.bag_configFile = '{ "foo": "bar" }';
       checks.bag_parse_commands.teardown.desc.should.equal('Delete databases, including documents');
       checks.bag_parse_commands.teardown.options.length.should.equal(3);
       checks.bag_parse_commands.teardown.action({
@@ -135,15 +123,14 @@ describe('cli', function () {
         dir: 'curr/dir/'
       });
       checks.couchpenter_url.should.equal('http://localhost:5984/somedb');
-      checks.couchpenter_dir.should.equal('curr/dir/');
-      checks.couchpenter_config.foo.should.equal('bar');
+      checks.couchpenter_opts.dir.should.equal('curr/dir/');
+      checks.couchpenter_configFile.should.equal('someconfigfile.js');
       checks.couchpenter_do_tasks.length.should.equal(1);
       checks.couchpenter_do_tasks[0].should.equal('tearDownDatabases');
       checks.couchpenter_do_exit.should.be.a('function');
     });
 
     it('should contain teardown-db command and delegate to couchpenter do when exec is called', function () {
-      mocks.bag_configFile = '{ "foo": "bar" }';
       checks.bag_parse_commands['teardown-db'].desc.should.equal('Alias for teardown');
       checks.bag_parse_commands['teardown-db'].options.length.should.equal(3);
       checks.bag_parse_commands['teardown-db'].action({
@@ -152,15 +139,14 @@ describe('cli', function () {
         dir: 'curr/dir/'
       });
       checks.couchpenter_url.should.equal('http://localhost:5984/somedb');
-      checks.couchpenter_dir.should.equal('curr/dir/');
-      checks.couchpenter_config.foo.should.equal('bar');
+      checks.couchpenter_opts.dir.should.equal('curr/dir/');
+      checks.couchpenter_configFile.should.equal('someconfigfile.js');
       checks.couchpenter_do_tasks.length.should.equal(1);
       checks.couchpenter_do_tasks[0].should.equal('tearDownDatabases');
       checks.couchpenter_do_exit.should.be.a('function');
     });
 
     it('should contain teardown-doc command and delegate to couchpenter do when exec is called', function () {
-      mocks.bag_configFile = '{ "foo": "bar" }';
       checks.bag_parse_commands['teardown-doc'].desc.should.equal('Delete documents only');
       checks.bag_parse_commands['teardown-doc'].options.length.should.equal(3);
       checks.bag_parse_commands['teardown-doc'].action({
@@ -169,15 +155,14 @@ describe('cli', function () {
         dir: 'curr/dir/'
       });
       checks.couchpenter_url.should.equal('http://localhost:5984/somedb');
-      checks.couchpenter_dir.should.equal('curr/dir/');
-      checks.couchpenter_config.foo.should.equal('bar');
+      checks.couchpenter_opts.dir.should.equal('curr/dir/');
+      checks.couchpenter_configFile.should.equal('someconfigfile.js');
       checks.couchpenter_do_tasks.length.should.equal(1);
       checks.couchpenter_do_tasks[0].should.equal('tearDownDocuments');
       checks.couchpenter_do_exit.should.be.a('function');
     });
 
     it('should contain reset command and delegate to couchpenter do when exec is called', function () {
-      mocks.bag_configFile = '{ "foo": "bar" }';
       checks.bag_parse_commands.reset.desc.should.equal('Delete then recreate databases and documents');
       checks.bag_parse_commands.reset.options.length.should.equal(3);
       checks.bag_parse_commands.reset.action({
@@ -186,8 +171,8 @@ describe('cli', function () {
         dir: 'curr/dir/'
       });
       checks.couchpenter_url.should.equal('http://localhost:5984/somedb');
-      checks.couchpenter_dir.should.equal('curr/dir/');
-      checks.couchpenter_config.foo.should.equal('bar');
+      checks.couchpenter_opts.dir.should.equal('curr/dir/');
+      checks.couchpenter_configFile.should.equal('someconfigfile.js');
       checks.couchpenter_do_tasks.length.should.equal(3);
       checks.couchpenter_do_tasks[0].should.equal('tearDownDatabases');
       checks.couchpenter_do_tasks[1].should.equal('setUpDatabases');
@@ -196,7 +181,6 @@ describe('cli', function () {
     });
 
     it('should contain reset-db command and delegate to couchpenter do when exec is called', function () {
-      mocks.bag_configFile = '{ "foo": "bar" }';
       checks.bag_parse_commands['reset-db'].desc.should.equal('Alias for reset');
       checks.bag_parse_commands['reset-db'].options.length.should.equal(3);
       checks.bag_parse_commands['reset-db'].action({
@@ -205,8 +189,8 @@ describe('cli', function () {
         dir: 'curr/dir/'
       });
       checks.couchpenter_url.should.equal('http://localhost:5984/somedb');
-      checks.couchpenter_dir.should.equal('curr/dir/');
-      checks.couchpenter_config.foo.should.equal('bar');
+      checks.couchpenter_opts.dir.should.equal('curr/dir/');
+      checks.couchpenter_configFile.should.equal('someconfigfile.js');
       checks.couchpenter_do_tasks.length.should.equal(3);
       checks.couchpenter_do_tasks[0].should.equal('tearDownDatabases');
       checks.couchpenter_do_tasks[1].should.equal('setUpDatabases');
@@ -215,7 +199,6 @@ describe('cli', function () {
     });
 
     it('should contain reset-doc command and delegate to couchpenter do when exec is called', function () {
-      mocks.bag_configFile = '{ "foo": "bar" }';
       checks.bag_parse_commands['reset-doc'].desc.should.equal('Delete then recreate documents only');
       checks.bag_parse_commands['reset-doc'].options.length.should.equal(3);
       checks.bag_parse_commands['reset-doc'].action({
@@ -224,49 +207,12 @@ describe('cli', function () {
         dir: 'curr/dir/'
       });
       checks.couchpenter_url.should.equal('http://localhost:5984/somedb');
-      checks.couchpenter_dir.should.equal('curr/dir/');
-      checks.couchpenter_config.foo.should.equal('bar');
+      checks.couchpenter_opts.dir.should.equal('curr/dir/');
+      checks.couchpenter_configFile.should.equal('someconfigfile.js');
       checks.couchpenter_do_tasks.length.should.equal(2);
       checks.couchpenter_do_tasks[0].should.equal('tearDownDocuments');
       checks.couchpenter_do_tasks[1].should.equal('setUpDocuments');
       checks.couchpenter_do_exit.should.be.a('function');
-    });
-
-    it('should use custom configuration file when specified and it exists', function () {
-      mocks.bag_configFile = '{ "foo": "bar" }';
-      try {
-        checks.bag_parse_commands.setup.action({
-          url: 'http://localhost:5984',
-          configFile: '../foobar.js'
-        });
-      } catch (e) {
-        should.fail('An error should not have been thrown since custom configuration should exist.' + e.message);
-      }
-      checks.bag_configFile.should.equal('../foobar.js');
-    });
-
-    it('should use default configuration file when configuration file is not specified', function () {
-      mocks.bag_configFile = '{ "foo": "bar" }';
-      try {
-        checks.bag_parse_commands.setup.action({
-          url: 'http://localhost:5984'
-        });
-      } catch (e) {
-        should.fail('An error should not have been thrown since custom configuration should exist.' + e.message);
-      }
-      checks.bag_configFile.should.equal('couchpenter.json');
-    });
-
-    it('should throw error when custom configuration file is specified but it does not exist', function () {
-      try {
-        checks.bag_parse_commands.setup.action({
-          url: 'http://localhost:5984',
-          configFile: '../somefilethatdoesnotexist.js'
-        });
-        should.fail('An error should have been thrown since custom configuration should not exist.');
-      } catch (e) {
-      }
-      checks.bag_configFile.should.equal('../somefilethatdoesnotexist.js');
     });
 
     it('should log task result', function () {
@@ -274,7 +220,6 @@ describe('cli', function () {
         'Created': ['foo', 'db1'],
         'Ignored': ['db2']
       }];
-      mocks.bag_configFile = '{ "foo": [], "db1": [], "db2": [] }';
       checks.bag_parse_commands.setup.action({
         url: 'http://localhost:5984/somedb',
         configFile: 'someconfigfile.js',